Fix enum/float arithmetic and uninitialized variable
- Fix enum/float compound assignment in CvGame.cpp:4507
- Fix enum/float arithmetic in CvUnit.cpp:14539
- Fix uninitialized variable in CvPlayer.cpp:43482

These fixes eliminate undefined behavior and type safety warnings
while maintaining C++03/TR1 compatibility.

+++ CvGameCoreDLL_Expansion2/CvGame.cpp
@@ -4504,7 +4504,7 @@ void CvGame::UpdateGameEra()
 	{
 		if (GET_TEAM((TeamTypes)iI).isAlive() && GET_TEAM((TeamTypes)iI).isMajorCiv())
 		{
-			fEra += GET_TEAM((TeamTypes)iI).GetCurrentEra();
+			fEra += static_cast<float>(GET_TEAM((TeamTypes)iI).GetCurrentEra());
 			iCount++;
 		}
 	}
